# docs/api-spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/maintenances": {
      "get": {
        "summary": "Get Maintenances",
        "operationId": "get_maintenances_maintenances_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MaintenanceItemResponse"
                  },
                  "title": "Response Get Maintenances Maintenances Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Maintenance",
        "operationId": "create_maintenance_maintenances_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaintenanceCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceItemResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/maintenances/by-ticket/{ticket_id}": {
      "get": {
        "summary": "Get Maintenance By Ticket",
        "operationId": "get_maintenance_by_ticket_maintenances_by_ticket__ticket_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Ticket Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceItemResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/maintenances/{maintenance_id}": {
      "get": {
        "summary": "Get Maintenance",
        "operationId": "get_maintenance_maintenances__maintenance_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "maintenance_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid7",
              "title": "Maintenance Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceItemResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update Maintenance",
        "operationId": "update_maintenance_maintenances__maintenance_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "maintenance_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid7",
              "title": "Maintenance Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_update_maintenance_maintenances__maintenance_id__patch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceItemResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Del Maintenance",
        "operationId": "del_maintenance_maintenances__maintenance_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "maintenance_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid7",
              "title": "Maintenance Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/maintenances/{maintenance_id}/pause": {
      "patch": {
        "summary": "Pause",
        "operationId": "pause_maintenances__maintenance_id__pause_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "maintenance_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid7",
              "title": "Maintenance Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_pause_maintenances__maintenance_id__pause_patch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketItemResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_pause_maintenances__maintenance_id__pause_patch": {
        "properties": {
          "payload": {
            "type": "string",
            "title": "Payload"
          },
          "initial_photo_file": {
            "anyOf": [
              {
                "type": "string",
                "format": "binary"
              },
              {
                "type": "null"
              }
            ],
            "title": "Initial Photo File"
          }
        },
        "type": "object",
        "required": [
          "payload"
        ],
        "title": "Body_pause_maintenances__maintenance_id__pause_patch"
      },
      "Body_update_maintenance_maintenances__maintenance_id__patch": {
        "properties": {
          "payload": {
            "type": "string",
            "title": "Payload"
          },
          "initial_photo_file": {
            "anyOf": [
              {
                "type": "string",
                "format": "binary"
              },
              {
                "type": "null"
              }
            ],
            "title": "Initial Photo File"
          }
        },
        "type": "object",
        "required": [
          "payload"
        ],
        "title": "Body_update_maintenance_maintenances__maintenance_id__patch"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MaintenanceCreate": {
        "properties": {
          "ticket_id": {
            "type": "integer",
            "title": "Ticket Id"
          },
          "maintenance_date": {
            "type": "string",
            "format": "date-time",
            "title": "Maintenance Date"
          }
        },
        "type": "object",
        "required": [
          "ticket_id",
          "maintenance_date"
        ],
        "title": "MaintenanceCreate"
      },
      "MaintenanceItemResponse": {
        "properties": {
          "maintenance_id": {
            "type": "string",
            "format": "uuid7",
            "title": "Maintenance Id"
          },
          "ticket_id": {
            "type": "integer",
            "title": "Ticket Id"
          },
          "maintenance_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Maintenance Date"
          },
          "maintenance_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Maintenance Description"
          },
          "spares": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SpareOut"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spares",
            "default": []
          },
          "technicians": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/TechnicianOut"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Technicians",
            "default": []
          },
          "initial_photo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Initial Photo Url"
          },
          "pdf_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pdf Url"
          },
          "photos": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/PhotoOut"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photos",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "maintenance_id",
          "ticket_id"
        ],
        "title": "MaintenanceItemResponse"
      },
...
      "PhotoOut": {
        "properties": {
          "photo_id": {
            "type": "integer",
            "title": "Photo Id"
          },
          "photo_url": {
            "type": "string",
            "title": "Photo Url"
          }
        },
        "type": "object",
        "required": [
          "photo_id",
          "photo_url"
        ],
        "title": "PhotoOut"
      },
...
      "SpareOut": {
        "properties": {
          "spare_id": {
            "type": "integer",
            "title": "Spare Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "price": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Price"
          },
          "qty": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Qty"
          }
        },
        "type": "object",
        "required": [
          "spare_id",
          "name",
          "price",
          "qty"
        ],
        "title": "SpareOut"
      },
...
      "TechnicianOut": {
        "properties": {
          "technician_id": {
            "type": "integer",
            "title": "Technician Id"
          },
          "technician_name": {
            "type": "string",
            "title": "Technician Name"
          },
          "start_hour": {
            "type": "string",
            "format": "time",
            "title": "Start Hour"
          },
          "end_hour": {
            "type": "string",
            "format": "time",
            "title": "End Hour"
          }
        },
        "type": "object",
        "required": [
          "technician_id",
          "technician_name",
          "start_hour",
          "end_hour"
        ],
        "title": "TechnicianOut"
      },
...
      "TicketItemResponse": {
        "properties": {
          "ticket_id": {
            "type": "integer",
            "title": "Ticket Id"
          },
          "ticket_date": {
            "type": "string",
            "format": "date",
            "title": "Ticket Date"
          },
          "ticket_description": {
            "type": "string",
            "title": "Ticket Description"
          },
          "priority": {
            "type": "string",
            "title": "Priority"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "market_name": {
            "type": "string",
            "title": "Market Name"
          },
          "market_city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market City"
          },
          "equipment_name": {
            "type": "string",
            "title": "Equipment Name"
          },
          "assigned_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned Name"
          }
        },
        "type": "object",
        "required": [
          "ticket_id",
          "ticket_date",
          "ticket_description",
          "priority",
          "status",
          "market_name",
          "equipment_name"
        ],
        "title": "TicketItemResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}